+++ shacl-diagram/src/main/java/fr/sparna/rdf/shacl/diagram/PlantUmlProperty.java
@@ -20,7 +20,7 @@ public class PlantUmlProperty {
 	protected PlantUmlBox value_node;
 	// contains a shortForm of the sh:class
 	protected String value_class_property;
-	protected Integer value_order_shacl;
+	protected Double value_order_shacl;
 	protected String value_hasValue;
 	protected PlantUmlBox value_qualifiedvalueshape;
 	protected String value_qualifiedMaxMinCount;
@@ -169,11 +169,11 @@ public class PlantUmlProperty {
 		this.value_class_property = value_class_property;
 	}
 
-	public Integer getValue_order_shacl() {
+	public Double getValue_order_shacl() {
 		return value_order_shacl;
 	}
 
-	public void setValue_order_shacl(Integer value_order_shacl) {
+	public void setValue_order_shacl(Double value_order_shacl) {
 		this.value_order_shacl = value_order_shacl;
 	}
